# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Beep communities openapi",
    "description": "API documentation for the Communities service",
    "contact": {
      "name": "communities-core@beep.ovh"
    },
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.0.1"
  },
  "paths": {
...
    "/friend-invitations": {
      "get": {
        "tags": [
          "friend-invitations"
        ],
        "operationId": "get_friend_invitations",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of friend invitations retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_FriendRequest"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/friend-requests": {
      "get": {
        "tags": [
          "friend-requests"
        ],
        "operationId": "get_friend_requests",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of friend requests retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_FriendRequest"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "friend-requests"
        ],
        "operationId": "create_friend_request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFriendRequestInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Friend request created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendRequest"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Friend request already exists"
          }
        }
      }
    },
    "/friend-requests/accept": {
      "post": {
        "tags": [
          "friend-requests"
        ],
        "operationId": "accept_friend_request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptFriendRequestInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Friend request accepted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Friend"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Friend request not found"
          }
        }
      }
    },
    "/friend-requests/decline": {
      "post": {
        "tags": [
          "friend-requests"
        ],
        "operationId": "decline_friend_request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeclineFriendRequestInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Friend request declined successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendRequest"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Friend request not found"
          }
        }
      }
    },
    "/friend-requests/{user_id_invited}": {
      "delete": {
        "tags": [
          "friend-requests"
        ],
        "operationId": "delete_friend_request",
        "parameters": [
          {
            "name": "user_id_invited",
            "in": "path",
            "description": "ID of the invited user whose friend request to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Friend request deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Friend request not found"
          }
        }
      }
    },
    "/friends": {
      "get": {
        "tags": [
          "friends"
        ],
        "operationId": "get_friends",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of friends retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_Friend"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/friends/{friend_id}": {
      "delete": {
        "tags": [
          "friends"
        ],
        "operationId": "delete_friend",
        "parameters": [
          {
            "name": "friend_id",
            "in": "path",
            "description": "ID of the friend to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Friend deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Friend not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcceptFriendRequestInput": {
        "type": "object",
        "required": [
          "user_id_requested"
        ],
        "properties": {
          "user_id_requested": {
            "$ref": "#/components/schemas/UserId"
          }
        }
      },
...
      "CreateFriendRequestInput": {
        "type": "object",
        "required": [
          "user_pseudo_invited"
        ],
        "properties": {
          "user_pseudo_invited": {
            "type": "string"
          }
        }
      },
...
      "DeclineFriendRequestInput": {
        "type": "object",
        "required": [
          "user_id_requested"
        ],
        "properties": {
          "user_id_requested": {
            "$ref": "#/components/schemas/UserId"
          }
        }
      },
...
      "Friend": {
        "type": "object",
        "required": [
          "user_id_1",
          "user_id_2",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "user_id_1": {
            "$ref": "#/components/schemas/UserId"
          },
          "user_id_2": {
            "$ref": "#/components/schemas/UserId"
          }
        }
      },
      "FriendRequest": {
        "type": "object",
        "required": [
          "user_id_requested",
          "user_id_invited",
          "status",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "user_id_invited": {
            "$ref": "#/components/schemas/UserId"
          },
          "user_id_requested": {
            "$ref": "#/components/schemas/UserId"
          }
        }
      },
...
      "PaginatedResponse_Friend": {
        "type": "object",
        "required": [
          "data",
          "total",
          "page"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "user_id_1",
                "user_id_2",
                "created_at"
              ],
              "properties": {
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "user_id_1": {
                  "$ref": "#/components/schemas/UserId"
                },
                "user_id_2": {
                  "$ref": "#/components/schemas/UserId"
                }
              }
            }
          },
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "total": {
            "$ref": "#/components/schemas/u64"
          }
        }
      },
      "PaginatedResponse_FriendRequest": {
        "type": "object",
        "required": [
          "data",
          "total",
          "page"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "user_id_requested",
                "user_id_invited",
                "status",
                "created_at"
              ],
              "properties": {
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "status": {
                  "type": "integer",
                  "format": "int32"
                },
                "user_id_invited": {
                  "$ref": "#/components/schemas/UserId"
                },
                "user_id_requested": {
                  "$ref": "#/components/schemas/UserId"
                }
              }
            }
          },
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "total": {
            "$ref": "#/components/schemas/u64"
          }
        }
      },
...
      "UserId": {
        "type": "string"
      },
      "u64": {
        "type": "integer",
        "format": "int64",
        "minimum": 0
      }
    }
  }
}